hooked up tokenizer logic
